Reduced Module number, changed extrapolation
I removed the feedback and self-evaluation modules, as they were not being triggered by the system and were kind of the same thing as the refining module. As such, their content was combined with the refining module.

In terms of architecture. I also changed how the agents extrapolate. Now, instead of the extrapolations being what each agent would say, the extrapolations are arguments for why the approach of each module would be advantageous in the given context. This gets the system to pick modules better but after the first ~7 messages, the extrapolations begin to break down as context is added. This could be solved by giving them their own independent histories, allowing their past responses to reinforce their future responses in the same way the judge works.