Parallel requests working
This is the version in which the parallel requests actually work. In order for this to happen, I had to create a new class that goes over standardConversation to mirror its functions in an async fassion. I also had to design and distribute a system-wide token counter to prevent hitting a rate limit now that the system is much faster. This new class is in the connversationTools module